[refactor] verifyFirebase.js 這個  middleware 拆成 verifyFirebase.js 與 verifyDB.js 兩個 middleware 做職責分離
