+++ server/src/middlewares/verifyFirebase.js
@@ -23,9 +23,7 @@
 //////////////////////////////////////////
 // V: 測試
 import admin from "../config/firebase.js";
-import db from "../config/db.js"; // 你自己的 Drizzle 初始化
-import { usersTable } from "../models/schema.js";
-import { eq } from "drizzle-orm";
+
 
 export async function verifyFirebase(req, res, next) {
   const authHeader = req.headers.authorization;
@@ -38,33 +36,11 @@ export async function verifyFirebase(req, res, next) {
 
   try {
     const decoded = await admin.auth().verifyIdToken(idToken);
-    const firebaseUid = decoded.uid;
-
-    // 查找 DB 是否已有對應 user
-    const result = await db
-      .select()
-      .from(usersTable)
-      .where(eq(usersTable.id, firebaseUid));
-    const user = result[0];
-
-    // 如果沒有，就創建一筆（首次登入）
-    if (!user) {
-      await db.insert(usersTable).values({
-        id: firebaseUid,
-        email: decoded.email,
-        username:
-          decoded.name || decoded.email?.split("@")[0] || "FirebaseUser",
-        password_hash: "firebase", // 填預設值，因為用不到
-        display_name: decoded.name || null,
-        // profile_image: decoded.picture || null,
-        bio: "", // 預設值
-      });
-    }
 
     // 加入 request，後續 API 可取得
+    req.userId = decoded.uid;
     req.firebaseUser = decoded;
-    req.userId = firebaseUid;
-
+    console.log(req.firebaseUser)
     next();
   } catch (err) {
     console.error("驗證失敗:", err);
