+++ server/src/routes/me.js
@@ -1,12 +1,12 @@
 // 測試用路由
 // routes/me.js
 import express from "express";
-import { verifyFirebase } from "../middlewares/varifyFirebase.js";
-
+import { verifyFirebase } from "../middlewares/verifyFirebase.js";
+import { verifyDB } from "../middlewares/verifyDB.js";
 const router = express.Router();
 
 // 測試用：驗證登入狀態與資料庫是否有寫入
-router.get("/", verifyFirebase, async (req, res) => {
+router.get("/", verifyFirebase, verifyDB, async (req, res) => {
 	res.json({
 		message: "Firebase token 驗證成功",
 		userId: req.userId,
